refactor: Move widget Databox into new file data_box.dart
