+++ lib/presentation/pages/detail/widgets/data_box.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class DataBox extends StatelessWidget {
+  const DataBox({super.key, required this.value, required this.title});
+
+  final dynamic value;
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(right: 10),
+      child: Container(
+        padding: const EdgeInsets.symmetric(horizontal: 28),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(10),
+          border: Border.all(color: Colors.white),
+        ),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              '$value',
+              style: const TextStyle(fontSize: 16, color: Colors.white),
+            ),
+            Text(
+              title,
+              style: const TextStyle(
+                fontSize: 16,
+                color: Colors.white,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
